requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-telegram-bot==22.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
